requirements.txt
alembic==1.13.1
fastapi==0.111.0
uvicorn[standard]
psycopg2==2.9.1
pydantic==2.7.1
starlette==0.37.2
typing-extensions==4.11.0
Python-dotenv==1.0.1
PyYAML==6.0.1
websockets==12.0
passlib==1.7.4
email-validator==2.1.1
argon2-cffi==23.1.0
click==8.1.7
SQLAlchemy==2.0.0
python-jose==3.3.0
pytest-asyncio==0.23.7
pytest-mock==3.14.0
httpcore==0.15.0
httpx==0.23.0








